requirements.txt
google-cloud-aiplatform[agent_engines,adk]
pydantic==2.11.7
pydantic-settings==2.7.1
cloudpickle==3.1.1
psycopg2==2.9.10
openinference-instrumentation-google-adk
google-adk
arize-phoenix-otel
sse-starlette
supabase
aikar